argon2: Clean up leftovers, migrate tests and add todo
